Redis: Extract AutcompleteIndexer module
